+++ __test__/ShiftTrackingProvider.test.js
@@ -287,27 +287,7 @@ describe('ShiftTrackingProvider', () => {
 		});
 	});
 
-	it('should render pausedShiftComponent when shiftStatus is paused and callback is provided', async () => {
-		isAuthorizedToUseStaffMS.mockResolvedValueOnce(true);
-		mockMMKV.getString.mockImplementation((key) => {
-			if (key === 'shift.status') return 'paused';
-			return null;
-		});
-
-		const {getByTestId} = render(
-			<ShiftTrackingProvider
-				pausedShiftComponent={<div data-testid="paused-component">Shift Paused</div>}>
-				<div>Normal Child</div>
-			</ShiftTrackingProvider>
-		);
-
-		await waitFor(() => {
-			expect(getByTestId('paused-component')).toBeDefined();
-		});
-	});
-
-	it('should render children normally when shiftStatus is not paused', async () => {
-		isAuthorizedToUseStaffMS.mockResolvedValueOnce(true);
+	it('should render children normally', async () => {
 		const {getByText} = render(
 			<ShiftTrackingProvider>
 				<div>Normal Child</div>