+++ setupTest/jest.setup.js
@@ -83,8 +83,20 @@ jest.mock('../lib/ShiftWorklogs', () => ({
 	default: {
 		open: jest.fn(),
 		finish: jest.fn(),
-		prepareWorkLogTypes: jest.fn(),
+		getShiftTrackedWorkLogs: jest.fn(),
 		getList: jest.fn(),
+	},
+}));
+
+// Mock Formatter
+jest.mock('../lib/Formatter', () => ({
+	__esModule: true,
+	default: {
+		formatShiftActivities: jest.fn(),
+		formatWorkLogTypes: jest.fn(),
+		formatWorkLogId: jest.fn(),
+		formatWorkLogsFromJanis: jest.fn(),
+		splitWorkLogsByStatus: jest.fn(),
 	},
 }));
 
@@ -92,6 +104,8 @@ jest.mock('../lib/ShiftWorklogs', () => ({
 jest.mock('../lib/db/TimeTrackerService', () => ({
 	addEvent: jest.fn(),
 	deleteAllEvents: jest.fn(),
+	getElapsedTime: jest.fn(),
+	searchEventByQuery: jest.fn(),
 }));
 
 // Mock StorageService
@@ -141,6 +155,21 @@ jest.mock('../lib/utils/helpers', () => {
 		promiseWrapper: jest.fn(),
 	};
 });
+
+jest.mock('../lib/utils/storage', () => ({
+	__esModule: true,
+	getShiftData: jest.fn(),
+}));
+
+// Mock TrackerRecords
+jest.mock('../lib/TrackerRecords', () => ({
+	__esModule: true,
+	default: {
+		getWorkLogsFromTimeTracker: jest.fn(),
+		getStartDateById: jest.fn(),
+		getEndDateById: jest.fn(),
+	},
+}));
 
 // Mock utils/provider functions
 jest.mock('../lib/utils/provider', () => ({